backend/requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0,<4.0.0
djangorestframework-simplejwt>=5.3.0,<6.0.0
django-cors-headers>=4.0.0,<5.0.0
psycopg2-binary>=2.9.0,<3.0.0
gunicorn>=20.1.0,<21.0.0
Pillow>=10.0.0,<11.0.0
django-filter>=23.0,<24.0
drf-yasg>=1.21.0,<2.0.0
